complete re-write of HexFish
Previously __lock_proc has nothing to do with locks, but prevents a message being fired in a hook from being processed in a hook, this is now done via decorators in HexFishCommand.

Topics support is temporarily removed as previously support was partial (only supporting the 332 server event when joining the channel, with no support for setting the topic or refreshing the topic upon any changes).